+++ imod/mf6/multimodel/exchange_creator.py
@@ -6,7 +6,6 @@ import pandas as pd
 import xarray as xr
 
 from imod.mf6.gwfgwf import GWFGWF
-from imod.mf6.gwfgwt import GWFGWT
 from imod.mf6.multimodel.modelsplitter import PartitionInfo
 from imod.mf6.utilities.grid import get_active_domain_slice, to_cell_idx
 from imod.typing import GridDataArray
